## Improve pass through endpoint latency to ~50ms (before PR was 400ms) 


### Before this 
<img width="868" alt="Screenshot 2024-11-22 at 5 11 49 PM" src="https://github.com/user-attachments/assets/64aa54df-5795-4b12-ba2a-2ebdbecf4c4a">

### Perf after this PR
<img width="1259" alt="Screenshot 2024-11-22 at 5 11 05 PM" src="https://github.com/user-attachments/assets/ee5366c4-ebf2-4bfc-849d-9fb13b317695">






<!-- e.g. "Implement user authentication feature" -->

## Relevant issues

<!-- e.g. "Fixes #000" -->

## Type

<!-- Select the type of Pull Request -->
<!-- Keep only the necessary ones -->

🆕 New Feature
🐛 Bug Fix
🧹 Refactoring
📖 Documentation
🚄 Infrastructure
✅ Test

## Changes

<!-- List of changes -->

## [REQUIRED] Testing - Attach a screenshot of any new tests passing locall
If UI changes, send a screenshot/GIF of working UI fixes

<!-- Test procedure -->

<!-- Korbit AI PR Description Start -->
## Description by Korbit AI

### What change is being made?

Improve pass-through endpoint latency to approximately 50ms by refactoring the HTTP client creation process to use a specialized async HTTPX client via the introduction of the `httpxSpecialProvider` enum and the `get_async_httpx_client` function.

### Why are these changes being made?

The change is aimed at optimizing the performance and latency of pass-through endpoint requests; previously, the generic HTTPX client was leading to a latency of around 400ms. By creating a custom async HTTP client specific to the pass-through endpoint, we reduce overhead and improve response times, also ensuring the consistent use of this function through tests that enforce it.

> Is this description stale? Ask me to generate a new description by commenting `/korbit-generate-pr-description`
<!-- Korbit AI PR Description End -->